+++ vblog-server-admin/src/main/java/com/hhoa/vblog/admin/service/impl/AmsArticleServiceImpl.java
@@ -39,7 +39,7 @@ public class AmsArticleServiceImpl implements AmsArticleService {
     }
 
     @Override
-    public void deleteArticle(Integer articleId) {
+    public void deleteArticle(Long articleId) {
         articleDao.deleteById(articleId);
     }
